pyproject.toml
# Guide (user-friendly):
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/

# Specification (technical, formal):
# https://packaging.python.org/en/latest/specifications/pyproject-toml/

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lumina-t2x"  # REQUIRED, is the only field that cannot be marked as dynamic.
version = "1.2.3"  # REQUIRED, although can be dynamic
description = "Lumina-T2X is a model for Text to Any Modality Generation"
readme = "README.md"

requires-python = ">=3.10"

license = {file = "LICENSE.txt"}

keywords = ["generation", "multi-modal", "transformer", "aigc", "diffusion"]

authors = [
  {name = "Alpha-VLLM", email = "author@example.com" }
]

maintainers = [
  {name = "Chris Liu", email = "author@example.com" },
  {name = "PommesPeter", email = "xepxa6823@gmail.com" }
]

classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
]

dependencies = [
  "diffusers",
  "fairscale",
  "accelerate",
  "tensorboard",
  "transformers",
  "gradio",
  "torchdiffeq",
  "click"
]

[project.optional-dependencies]
dev = ["coverage", "pre-commit", "isort", "black"]
image = ["diffusers", "fairscale", "accelerate", "tensorboard", "transformers", "gradio", "torchdiffeq", "click"
]
music = ["soundfile", "omegaconf", "torchdyn", "pytorch_lightning", "pytorch_memlab", "einops", "ninja", "torchlibrosa", "protobuf", "sentencepiece", "gradio", "transformers"]
audio = ["soundfile", "omegaconf", "torchdyn", "pytorch_lightning", "pytorch_memlab", "einops", "ninja", "torchlibrosa", "protobuf", "sentencepiece", "gradio", "transformers"]


[project.scripts]
lumina = "lumina_t2i:entry_point"
lumina_next = "lumina_next_t2i:entry_point"

[tool.setuptools.packages.find]
exclude = ["assets*"]

[tool.wheel]
exclude = ["assets*"]
